+++ src/mainframe.h
@@ -156,7 +156,7 @@ public:
 
     void FindItemName(Node* node);
 
-    void setStatusText(const ttlib::cstr& txt, size_t pane = 1);
+    void setStatusText(const ttlib::cstr& txt, int pane = 1);
     wxStatusBar* OnCreateStatusBar(int number, long style, wxWindowID id, const wxString& name) override;
 
     bool SaveWarning();
